Refactor presentational data methods to decorator and simplify implementation
